extracted the router to a different file and working on post request
