+++ assets/templates/index.html
@@ -1,6 +1,31 @@
-{{ template "header" . }}
+<html lang="en">
+<head>
+    <title>Kalenderapplikation</title>
+    <!-- Bootstrap CSS -->
+    <meta charset="utf-8">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.5.0/font/bootstrap-icons.css">
+</head>
+<body>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
 <div class="container-fluid">
-
+    <div class="row">
+        <div class="column" style="width: 33%"></div>
+        <div class="column" style="width: 33%">
+            <form method="post">
+                <div class="form-group">
+                    <label for="user">Username:</label>
+                    <input type="text" class="form-control" id="user" placeholder="User eingeben">
+                </div>
+                <div class="form-group">
+                    <label for="password">Passwort</label>
+                    <input type="password" class="form-control" id="password" placeholder="Passwort">
+                </div>
+                <button class="btn btn-primary" type="submit" style="margin-top: 20px">Einloggen</button>
+            </form>
+        </div>
+        <div class="column" style="width: 33%"></div>
+    </div>
 </div>
 {{ template "footer" . }}
